+++ src/events/processor/lwtProcessor.js
@@ -1,7 +1,5 @@
 import logger from '../../common/logger';
-import SonoffPowTopicHanlders from '../../services/sonoffPowTopicHandler';
 
-const topicHanlders = SonoffPowTopicHanlders();
 const STATUS_OFFLINE = 'Offline';
 const STATUS_ONLINE = 'Online';
 
@@ -9,17 +7,21 @@ const LwtProcessor = providers => ({
   process: (event) => {
     logger.log('info', `lwt processor ${JSON.stringify(event)}`);
 
-    const deviceName = topicHanlders.extractNameFromTopic('tele', event.Payload.Topic);
-
-    providers.deviceProvider.findByName(deviceName).then((dev) => {
-      if (event.Payload.Status === STATUS_OFFLINE) {
-        providers.deviceProvider.updateByDeviceId(dev.deviceId,
-          { ...dev, status: { online: false } });
-      } else if (event.Payload.Status === STATUS_ONLINE) {
-        providers.deviceProvider.updateByDeviceId(dev.deviceId,
-          { ...dev, status: { online: true } });
-      }
-    }).catch(err => logger.log('error', err));
+    return new Promise((resolve, reject) => {
+      providers.deviceProvider.findByDeviceId(event.Payload.DeviceId).then((dev) => {
+        if (event.Payload.Status === STATUS_OFFLINE) {
+          providers.deviceProvider
+            .updateByDeviceId(dev.deviceId,
+              { ...dev, status: { online: false } })
+            .then(() => resolve());
+        } else if (event.Payload.Status === STATUS_ONLINE) {
+          providers.deviceProvider
+            .updateByDeviceId(dev.deviceId,
+              { ...dev, status: { online: true } })
+            .then(() => resolve());
+        }
+      }).catch(err => reject(err));
+    });
   },
 });