+++ src/events/processor/powerStateProcessor.js
@@ -12,23 +12,27 @@ const PowerStateProcessor = (providers, pnub) => ({
   process: (event) => {
     logger.log('info', `power action processor ${JSON.stringify(event)}`);
 
-    providers
-      .deviceProvider
-      .findByDeviceId(event.deviceId)
-      .then((dev) => {
-        if (dev.commands && dev.commands.power) {
-          pnub.publish(event.gateway, {
-            type: 'MQTT',
-            payload: {
-              topic: dev.commands.power.replace('mqtt:', ''),
-              value: event.param,
-            },
-          });
-        } else {
-          logger.log('error', `Device ${event.deviceId} doesn't have power command configured.`);
-        }
-      })
-      .catch(err => logger.log('error', err));
+    return new Promise((resolve, reject) => {
+      providers
+        .deviceProvider
+        .findByDeviceId(event.deviceId)
+        .then((dev) => {
+          if (dev.commands && dev.commands.power) {
+            pnub.publish(event.gateway, {
+              type: 'MQTT',
+              payload: {
+                topic: dev.commands.power.replace('mqtt:', ''),
+                value: event.param,
+              },
+            });
+            resolve();
+          } else {
+            const errMessage = `Device ${event.deviceId} doesn't have power command configured.`;
+            reject(new Error(errMessage));
+          }
+        })
+        .catch(err => reject(err));
+    });
   },
 });